s/sys_arc/requirements.txt
pulumi>=3.0.0,<4.0.0
pulumi-gcp>=7.0.0,<8.0.0
google-cloud-bigquery
functions_framework

